Update and rename kali_x64.sh to kali_general.sh

+++ kali_general.sh
@@ -49,7 +49,7 @@ make -j$(nproc)
 sudo make altinstall
 
 # Install pwncat-cs
-pipx install pwncat-cs --python python3.11
+pipx install git+https://github.com/Chocapikk/pwncat-vl --python python3.11
 
 # Install uvx for installing subsurfer
 curl -LsSf https://astral.sh/uv/install.sh | sh
@@ -65,38 +65,55 @@ cargo build --release
 sudo cp ~/RustScan/target/release/rustscan /usr/bin/rustscan
 sudo chmod +x /usr/bin/rustscan
 
-# Install volatility2
-sudo apt install -y python2 python2.7-dev libpython2-dev
-curl https://bootstrap.pypa.io/pip/2.7/get-pip.py --output get-pip.py
-sudo python2 get-pip.py
-sudo python2 -m pip install -U setuptools wheel
 
-python2 -m pip install -U  yara pycrypto pillow openpyxl ujson pytz ipython capstone
-pip2 install distorm3==3.3.4
-sudo python2 -m pip install yara
-sudo ln -s /usr/local/lib/python2.7/dist-packages/usr/lib/libyara.so /usr/lib/libyara.so
-python2 -m pip install -U git+https://github.com/volatilityfoundation/volatility.git
+# Install uv
+curl -LsSf https://astral.sh/uv/install.sh | sh
 
 
+# Install enum4linux-ng
+uv tool install git+https://github.com/cddmp/enum4linux-ng
 
-# install volatility3
-sudo apt install -y python3 python3-dev libpython3-dev python3-pip python3-setuptools python3-wheel
+###################################
+# x86_64 specific installations
+###################################
 
-python3 -m pip install -U distorm3 yara pycrypto pillow openpyxl ujson pytz ipython capstone
-python3 -m pip install -U git+https://github.com/volatilityfoundation/volatility3.git
+if [ "$ARCH" = "x86_64" ]; then
+    # Install volatility2
+    sudo apt install -y python2 python2.7-dev libpython2-dev
+    curl https://bootstrap.pypa.io/pip/2.7/get-pip.py --output get-pip.py
+    sudo python2 get-pip.py
+    sudo python2 -m pip install -U setuptools wheel
 
-echo "export PATH=/home/`whoami`/.local/bin:$PATH\n" >> ~/.bashrc
-. ~/.bashrc
+    python2 -m pip install -U  yara pycrypto pillow openpyxl ujson pytz ipython capstone
+    pip2 install distorm3==3.3.4
+    sudo python2 -m pip install yara
+    sudo ln -s /usr/local/lib/python2.7/dist-packages/usr/lib/libyara.so /usr/lib/libyara.so
+    python2 -m pip install -U git+https://github.com/volatilityfoundation/volatility.git
 
 
 
-# install mullvad ~lenny~
-wget -O mullvad_vpn.deb https://mullvad.net/en/download/app/deb/latest
-sudo dpkg -i mullvad_vpn.deb
-rm -rf mullvad_vpn.deb
+    # install volatility3
+    sudo apt install -y python3 python3-dev libpython3-dev python3-pip python3-setuptools python3-wheel
 
-wget -O mullvad.tar.gz https://mullvad.net/en/download/browser/linux-x86_64/latest
-tar -xvf mullvad.tar.gz
+    python3 -m pip install -U distorm3 yara pycrypto pillow openpyxl ujson pytz ipython capstone
+    python3 -m pip install -U git+https://github.com/volatilityfoundation/volatility3.git
+
+    echo "export PATH=/home/`whoami`/.local/bin:$PATH\n" >> ~/.bashrc
+    . ~/.bashrc
+
+    # install mullvad ~lenny~
+    wget -O mullvad_vpn.deb https://mullvad.net/en/download/app/deb/latest
+    sudo dpkg -i mullvad_vpn.deb
+    rm -rf mullvad_vpn.deb
+
+    wget -O mullvad.tar.gz https://mullvad.net/en/download/browser/linux-x86_64/latest
+    tar -xvf mullvad.tar.gz
+
+elif [ "$ARCH" = "aarch64" ]; then
+    # Install vmtools
+    sudo apt install -y --reinstall open-vm-tools-desktop fuse
+
+else
 
 
 # Install custom scripts
@@ -111,6 +128,5 @@ chmod +x ~/scripts/feroxbuster.sh
 # install bloodhound (kali-only)
 sudo apt-get install -y bloodhound
 sudo bloodhound-setup
-
